+++ src/components/WYSIWYGInput/WYSIWYGInput.stories.tsx
@@ -1,6 +1,6 @@
 // also exported from '@storybook/react' if you can deal with breaking changes in 6.1
 import { Meta, Story } from '@storybook/react/types-6-0';
-import React, { useRef, useState } from 'react';
+import React, { useEffect, useRef, useState } from 'react';
 import { WYSIWYGInput, WYSIWYGInputProps, WYSIWYGInputRef } from '.';
 import { Button } from '../Button';
 import { HTMLParsedContent } from '../HTMLParsedContent';
@@ -38,6 +38,7 @@ export default {
       }
     },
     onChange: { action: 'onChange' },
+    onHtmlChangeSlow: { action: 'onHtmlChangeSlow' },
     onBlur: { action: 'onBlur' },
     onFocus: { action: 'onFocus' }
   }
@@ -51,8 +52,7 @@ export default {
  * @param props.className the component classes
  */
 const Template: Story<WYSIWYGInputProps & { content?: string; className?: string }> = ({ ...props }) => (
-  // Clear the onHtmlChangeSlow event that storybook provides automatically because it's slow
-  <WYSIWYGInput {...props} onHtmlChangeSlow={undefined} />
+  <WYSIWYGInput {...props} />
 );
 
 export const WYSIWYG = Template.bind({});
@@ -70,6 +70,34 @@ WithError.args = {
 export const DefaultValue = Template.bind({});
 DefaultValue.args = {
   defaultHtmlValue:
+    '<p>Lorem&nbsp;ipsum&nbsp;dolor,&nbsp;sit&nbsp;<a href="https://google.com">amet</a>&nbsp;consectetur&nbsp;adipisicing&nbsp;elit.&nbsp;Atque&nbsp;quaerat,&nbsp;minus&nbsp;suscipit&nbsp;totam&nbsp;nam&nbsp;sequi&nbsp;hic&nbsp;expedita&nbsp;quae&nbsp;fugit&nbsp;blanditiis&nbsp;ad&nbsp;voluptate&nbsp;eligendi&nbsp;adipisci.&nbsp;Quaerat&nbsp;similique&nbsp;numquam&nbsp;corporis&nbsp;molestias,&nbsp;blanditiis&nbsp;enim&nbsp;saepe&nbsp;accusamus&nbsp;debitis&nbsp;ipsa&nbsp;provident&nbsp;cumque&nbsp;voluptatum&nbsp;dolor&nbsp;beatae&nbsp;odit,&nbsp;odio&nbsp;ipsum&nbsp;maiores!&nbsp;Reiciendis,&nbsp;debitis&nbsp;fugiat&nbsp;consequuntur&nbsp;ex&nbsp;quae&nbsp;necessitatibus&nbsp;laboriosam&nbsp;sit&nbsp;quas&nbsp;et&nbsp;nam!&nbsp;Maxime,&nbsp;necessitatibus&nbsp;velit&nbsp;veniam,&nbsp;assumenda&nbsp;minima&nbsp;soluta,&nbsp;molestias&nbsp;rem&nbsp;eos&nbsp;harum&nbsp;reiciendis&nbsp;tempore&nbsp;consequatur?&nbsp;Vel&nbsp;quos&nbsp;similique&nbsp;rem,&nbsp;delectus&nbsp;maiores&nbsp;inventore&nbsp;quisquam&nbsp;ab&nbsp;asperiores&nbsp;explicabo&nbsp;accusantium&nbsp;quas&nbsp;doloribus&nbsp;iusto&nbsp;dolorem&nbsp;repellat.&nbsp;Necessitatibus&nbsp;repellendus&nbsp;consectetur&nbsp;hic&nbsp;autem&nbsp;perspiciatis&nbsp;voluptatibus,&nbsp;eius&nbsp;laudantium&nbsp;quas&nbsp;iste,&nbsp;nobis&nbsp;unde?</p> <p></p> <p>Totam&nbsp;numquam&nbsp;sunt&nbsp;iure&nbsp;recusandae&nbsp;harum&nbsp;nihil&nbsp;itaque&nbsp;assumenda&nbsp;facilis&nbsp;in&nbsp;repellat&nbsp;sed&nbsp;sapiente,&nbsp;esse&nbsp;blanditiis&nbsp;fuga&nbsp;aliquid&nbsp;pariatur&nbsp;architecto&nbsp;laboriosam.&nbsp;Id&nbsp;ex&nbsp;temporibus&nbsp;voluptatum&nbsp;maiores,&nbsp;error&nbsp;eveniet&nbsp;totam&nbsp;in&nbsp;ipsa&nbsp;beatae&nbsp;iure&nbsp;obcaecati&nbsp;corrupti,&nbsp;numquam&nbsp;nesciunt&nbsp;ipsum&nbsp;a&nbsp;dicta&nbsp;ad!&nbsp;Veritatis&nbsp;labore&nbsp;cum&nbsp;necessitatibus,&nbsp;a&nbsp;delectus&nbsp;doloremque&nbsp;voluptate&nbsp;ut&nbsp;blanditiis&nbsp;nisi&nbsp;atque&nbsp;soluta&nbsp;repellendus&nbsp;earum&nbsp;fugiat&nbsp;fugit&nbsp;totam&nbsp;harum&nbsp;commodi&nbsp;saepe!&nbsp;Totam&nbsp;qui&nbsp;cumque&nbsp;recusandae&nbsp;beatae&nbsp;unde&nbsp;atque&nbsp;nobis&nbsp;ipsa&nbsp;nihil</p>'
+};
+
+/**
+ * A story that displays a WYSIWYGInput example
+ *
+ * @param props the story props
+ * @param props.content  the content property set on controls
+ * @param props.htmlValueSlow The html value to fill the input with
+ * @param props.className the component classes
+ */
+const ValueChangeTemplate: Story<WYSIWYGInputProps & { content?: string; className?: string }> = ({
+  htmlValueSlow,
+  ...props
+}) => {
+  const [htmlValue, setHtmlValue] = useState(htmlValueSlow);
+
+  useEffect(() => {
+    const id = setInterval(() => setHtmlValue((htmlValue) => htmlValue + ' a'), 1000);
+    return () => clearInterval(id);
+  }, []);
+
+  return <WYSIWYGInput {...props} htmlValueSlow={htmlValue} onHtmlChangeSlow={setHtmlValue} />;
+};
+
+export const ValueChange = ValueChangeTemplate.bind({});
+ValueChange.args = {
+  htmlValueSlow:
     '<p>Lorem&nbsp;ipsum&nbsp;dolor,&nbsp;sit&nbsp;<a href="https://google.com">amet</a>&nbsp;consectetur&nbsp;adipisicing&nbsp;elit.&nbsp;Atque&nbsp;quaerat,&nbsp;minus&nbsp;suscipit&nbsp;totam&nbsp;nam&nbsp;sequi&nbsp;hic&nbsp;expedita&nbsp;quae&nbsp;fugit&nbsp;blanditiis&nbsp;ad&nbsp;voluptate&nbsp;eligendi&nbsp;adipisci.&nbsp;Quaerat&nbsp;similique&nbsp;numquam&nbsp;corporis&nbsp;molestias,&nbsp;blanditiis&nbsp;enim&nbsp;saepe&nbsp;accusamus&nbsp;debitis&nbsp;ipsa&nbsp;provident&nbsp;cumque&nbsp;voluptatum&nbsp;dolor&nbsp;beatae&nbsp;odit,&nbsp;odio&nbsp;ipsum&nbsp;maiores!&nbsp;Reiciendis,&nbsp;debitis&nbsp;fugiat&nbsp;consequuntur&nbsp;ex&nbsp;quae&nbsp;necessitatibus&nbsp;laboriosam&nbsp;sit&nbsp;quas&nbsp;et&nbsp;nam!&nbsp;Maxime,&nbsp;necessitatibus&nbsp;velit&nbsp;veniam,&nbsp;assumenda&nbsp;minima&nbsp;soluta,&nbsp;molestias&nbsp;rem&nbsp;eos&nbsp;harum&nbsp;reiciendis&nbsp;tempore&nbsp;consequatur?&nbsp;Vel&nbsp;quos&nbsp;similique&nbsp;rem,&nbsp;delectus&nbsp;maiores&nbsp;inventore&nbsp;quisquam&nbsp;ab&nbsp;asperiores&nbsp;explicabo&nbsp;accusantium&nbsp;quas&nbsp;doloribus&nbsp;iusto&nbsp;dolorem&nbsp;repellat.&nbsp;Necessitatibus&nbsp;repellendus&nbsp;consectetur&nbsp;hic&nbsp;autem&nbsp;perspiciatis&nbsp;voluptatibus,&nbsp;eius&nbsp;laudantium&nbsp;quas&nbsp;iste,&nbsp;nobis&nbsp;unde?</p> <p></p> <p>Totam&nbsp;numquam&nbsp;sunt&nbsp;iure&nbsp;recusandae&nbsp;harum&nbsp;nihil&nbsp;itaque&nbsp;assumenda&nbsp;facilis&nbsp;in&nbsp;repellat&nbsp;sed&nbsp;sapiente,&nbsp;esse&nbsp;blanditiis&nbsp;fuga&nbsp;aliquid&nbsp;pariatur&nbsp;architecto&nbsp;laboriosam.&nbsp;Id&nbsp;ex&nbsp;temporibus&nbsp;voluptatum&nbsp;maiores,&nbsp;error&nbsp;eveniet&nbsp;totam&nbsp;in&nbsp;ipsa&nbsp;beatae&nbsp;iure&nbsp;obcaecati&nbsp;corrupti,&nbsp;numquam&nbsp;nesciunt&nbsp;ipsum&nbsp;a&nbsp;dicta&nbsp;ad!&nbsp;Veritatis&nbsp;labore&nbsp;cum&nbsp;necessitatibus,&nbsp;a&nbsp;delectus&nbsp;doloremque&nbsp;voluptate&nbsp;ut&nbsp;blanditiis&nbsp;nisi&nbsp;atque&nbsp;soluta&nbsp;repellendus&nbsp;earum&nbsp;fugiat&nbsp;fugit&nbsp;totam&nbsp;harum&nbsp;commodi&nbsp;saepe!&nbsp;Totam&nbsp;qui&nbsp;cumque&nbsp;recusandae&nbsp;beatae&nbsp;unde&nbsp;atque&nbsp;nobis&nbsp;ipsa&nbsp;nihil</p>'
 };